requirements.txt
streamlit>=1.35.0
requests>=2.31.0
google-generativeai>=0.7.0
gspread>=6.0.0
google-auth>=2.29.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
